requirements.txt
alabaster==0.7.12
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-debug-toolbar==3.2
django-extensions==3.1.1
djangorestframework==3.12.2
docutils==0.16
factory-boy==3.2.0
Faker==6.6.1
idna==2.10
imagesize==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
netaddr==0.8.0
packaging==20.9
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
